requirements.txt
Django==5.2.7
google-genai
